+++ messages/am/config.php
@@ -1,8 +1,10 @@
 <?php
 return [
   'An event has been detected. This configuration may be overridden by the event.' => 'አንድ ክስተት ተገኝቷል። ይህ ውቅር በክስተቱ ሊሻር ይችላል።',
-  'Banner content (HTML)' => 'የሰንደቅ ይዘት (ኤችቲኤምኤል)',
+  'Banner content for logged-in users (HTML)' => 'የሰንደቅ ይዘት ለገቡ ተጠቃሚዎች (ኤችቲኤምኤል)',
+  'Banner content for visitors / logged-out users (HTML)' => 'የባነር ይዘት ለጎብኚዎች / ለወጡ ተጠቃሚዎች (ኤችቲኤምኤል)',
   'Close button' => 'ዝጋ አዝራር',
   'Enabled' => 'ነቅቷል',
   'Examples of CSS values that can be overwritten:' => 'ሊገለበጡ የሚችሉ የCSS እሴቶች ምሳሌዎች፡-',
+  'If the content is empty, the banner will not be displayed.' => 'ይዘቱ ባዶ ከሆነ ባነር አይታይም።',
 ];