+++ docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+1.1.0 (March 29, 2025)
+--------------------
+- Enh: Add a content field for visitors
+- Chg: CSS variables: `--banner-height` -> `--hh-banner-height`, `--banner-font-color` -> `--hh-banner-font-color` and `--banner-bg-color` -> `--hh-banner-bg-color`
+
 1.0.2 (March 24, 2025)
 --------------------
 - Fix #2: Wrong links to the GitHub repository